replace interface
