+++ packages/client/src/lib/utils.ts
@@ -4,11 +4,3 @@ import { twMerge } from 'tailwind-merge'
 export function cn(...inputs: ClassValue[]) {
   return twMerge(clsx(inputs))
 }
-
-export function getCookie(name: string) {
-  const regex = new RegExp(`(^| )${name}=([^;]+)`)
-  const match = document.cookie.match(regex)
-  if (match) {
-    return match[2]
-  }
-}
